seikai_hack/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
requests==2.31.0
openai==1.3.7
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
PyPDF2==3.0.1
